ai_agent/vision/image_analyzer.py
"""Image analysis and description functionality."""

import asyncio
import logging
from typing import Optional
from PIL import Image
import base64
import io

from ..core.config import Config


class ImageAnalyzer:
    """Image analysis engine for describing images."""
    
    def __init__(self, config: Config):
        """Initialize the image analyzer."""
        self.config = config
        self.logger = logging.getLogger(__name__)
    
    async def describe_image(self, image_path: str) -> str:
        """Analyze an image and return a description."""
        try:
            # Load and validate the image
            image = await self._load_image(image_path)
            if not image:
                return "Could not load the image. Please check the file path and format."
            
            # For now, provide a basic analysis
            # In a full implementation, this would use AI vision models
            description = await self._analyze_basic_properties(image)
            
            return description
            
        except Exception as e:
            self.logger.error(f"Error analyzing image: {e}")
            return "I encountered an error while analyzing the image."
    
    async def _load_image(self, image_path: str) -> Optional[Image.Image]:
        """Load an image from file path."""
        try:
            loop = asyncio.get_event_loop()
            image = await loop.run_in_executor(None, Image.open, image_path)
            return image
            
        except Exception as e:
            self.logger.error(f"Error loading image: {e}")
            return None
    
    async def _analyze_basic_properties(self, image: Image.Image) -> str:
        """Analyze basic image properties."""
        try:
            width, height = image.size
            mode = image.mode
            format_name = image.format or "Unknown"
            
            # Basic color analysis
            if mode == "RGB":
                color_description = "a color image"
            elif mode == "L":
                color_description = "a grayscale image"
            elif mode == "RGBA":
                color_description = "a color image with transparency"
            else:
                color_description = f"an image in {mode} format"
            
            # Size category
            if width * height > 2000000:  # > 2MP
                size_description = "high resolution"
            elif width * height > 500000:  # > 0.5MP
                size_description = "medium resolution"
            else:
                size_description = "low resolution"
            
            description = f"This is {color_description} with dimensions {width} by {height} pixels. It's a {size_description} {format_name} image."
            
            # Add orientation
            if width > height * 1.3:
                description += " The image is in landscape orientation."
            elif height > width * 1.3:
                description += " The image is in portrait orientation."
            else:
                description += " The image is roughly square."
            
            # Note: In a full implementation, this would include:
            # - Object detection and recognition
            # - Scene analysis
            # - Text extraction (OCR)
            # - Facial recognition
            # - Color palette analysis
            # - Composition analysis
            
            description += "\n\nNote: This is a basic analysis. For detailed image description including objects, scenes, and text, AI vision models would be integrated in the full implementation."
            
            return description
            
        except Exception as e:
            self.logger.error(f"Error in basic analysis: {e}")
            return "Could not analyze the image properties."
    
    def supported_formats(self) -> list:
        """Get list of supported image formats."""
        return ["JPEG", "PNG", "GIF", "BMP", "TIFF", "WEBP"]