requirements.txt
# Core AI and ML libraries
openai==1.3.0
anthropic==0.8.0
transformers==4.35.0
torch==2.1.0

# Speech processing
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.11

# Web framework and API
fastapi==0.104.0
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Image processing
Pillow==10.0.1
opencv-python==4.8.1.78

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.0
pydantic==2.4.2
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0